mui changed a bit on single rec page but still needs work
