+++ src/components/SingleRecord.js
@@ -10,11 +10,19 @@ import Typography from "@mui/material/Typography";
 import { Box } from "@mui/system";
 import CreateReviewForm from "./CreateReviewForm";
 import SingleRecReviews from "./SingleRecReviews";
+import { useState } from "react";
+import Rating from "@mui/material/Rating";
 
 const SingleRecord = () => {
   const { selectedRecord, loadingRecord } = useSelector((state) => {
     return state.selectedRecord;
   });
+
+  const [showReviews, setShowReviews] = useState(false);
+
+  const displayAllReviews = () => {
+    setShowReviews(!showReviews);
+  };
 
   const params = useParams("");
   const recordId = params.id;
@@ -42,9 +50,14 @@ const SingleRecord = () => {
   //check if selected record exist
   if (!Object.keys(selectedRecord).length)
     return <h1>Oops, this record doesn't exist, please try again</h1>;
+  const sumOfRatings = selectedRecord.reviews.reduce(
+    (rating, nextRating) => rating + nextRating.reviewRating,
+    0
+  );
+  const avgRating = (sumOfRatings / selectedRecord.reviews.length).toFixed(1);
 
   return (
-    <Container maxWidth="100vw">
+    <Container maxWidth="sm">
       <Box>
         <img
           component="img"
@@ -55,6 +68,15 @@ const SingleRecord = () => {
         <Typography gutterBottom variant="h5" component="div">
           <h3>{selectedRecord.albumName}</h3>
         </Typography>
+        {selectedRecord.reviews.length > 0 && (
+          <>
+            <Typography component="legend">
+              <b>Average Rating</b>
+            </Typography>
+            <Rating name="read-only" value={Number(avgRating)} readOnly />
+          </>
+        )}
+        <br></br>
         <Typography variant="body2" color="text.secondary">
           <span>
             <b>Artist:</b> {selectedRecord.artist}
@@ -82,7 +104,16 @@ const SingleRecord = () => {
             ))}
           </span>
         </Typography>
-        <SingleRecReviews selectedRecord={selectedRecord} />
+        <br></br>
+        {showReviews ? (
+          <>
+            <Button onClick={displayAllReviews}>Hide reviews</Button>
+            <SingleRecReviews selectedRecord={selectedRecord} />
+          </>
+        ) : (
+          <Button onClick={displayAllReviews}>See all reviews</Button>
+        )}
+
         <Button size="small" href={"/"}>
           Back
         </Button>
